refactor: Make the hooks more modular
- Added `useChat` hook to manage chat state, including sessions and messages.
- Introduced `useMessageHandler` for sending and resending messages.
- Created `useChatSessions` for managing chat sessions (create, select, delete).
- Implemented `useChatStorage` for loading and saving chat data.
- Added `useChatScroll` to handle scrolling behavior in chat container.
- Introduced `useThemeManager` for managing dark/light mode.
- Created `WelcomeMessage` component to greet authenticated users.
- Updated `InputArea` to include a button for resending the last message.
- Added utility functions for generating chat IDs and creating empty token usage.
- Refactored local storage service to ensure valid token usage in chat sessions.
- Added tests for the `useChat` hook to verify functionality and theme toggling.
