+++ app/home/page-client.tsx
@@ -10,6 +10,8 @@ import { InputArea } from '@/components/input-area';
 import { useChat } from '@/hooks/use-chat';
 import { useState, useEffect } from 'react';
 import ThemeToggle from '@/components/theme-toggle';
+import WelcomeMessage from '@/components/welcome-message';
+import { MessageRole } from '@inference-gateway/sdk';
 
 export const dynamic = 'force-dynamic';
 
@@ -29,6 +31,7 @@ export default function PageClient() {
     handleSendMessage,
     handleSelectChat,
     handleDeleteChat,
+    handleResendLastMessage,
     chatContainerRef,
   } = useChat();
 
@@ -36,6 +39,8 @@ export default function PageClient() {
   const isMobile = useIsMobile();
   const [showSidebar, setShowSidebar] = useState(!isMobile);
   const [hasMessages, setHasMessages] = useState(messages.length > 0);
+
+  const hasUserMessages = messages.some(message => message.role === MessageRole.user);
 
   useEffect(() => {
     setShowSidebar(!isMobile);
@@ -102,40 +107,68 @@ export default function PageClient() {
         )}
       >
         {/* Header */}
-        <header className="border-b border-[hsl(var(--chat-sidebar-border))] bg-[hsl(var(--chat-background))] py-4 px-3.5 relative h-14">
-          {/* Left - Chat history button */}
-          <button
-            onClick={() => setShowSidebar(!showSidebar)}
-            className={cn(
-              'absolute left-3.5 top-1/2 -translate-y-1/2 flex items-center justify-center',
-              'w-7 h-7 rounded-md hover:bg-accent',
-              'text-muted-foreground hover:text-foreground transition-colors'
-            )}
-            aria-label="Toggle chat history"
-          >
-            <Menu className="h-5 w-5" />
-          </button>
+        <header
+          className={cn(
+            'border-b border-[hsl(var(--chat-sidebar-border))] bg-[hsl(var(--chat-background))] px-3.5 relative flex flex-col',
+            isMobile ? 'py-3 h-auto min-h-[4.5rem]' : 'py-4 h-14'
+          )}
+        >
+          {/* Top row - controls and welcome message */}
+          <div className="flex items-center justify-between w-full relative">
+            {/* Left - Chat history button and welcome message */}
+            <div className="flex items-center gap-2">
+              <button
+                onClick={() => setShowSidebar(!showSidebar)}
+                className={cn(
+                  'flex items-center justify-center',
+                  'w-7 h-7 rounded-md hover:bg-accent',
+                  'text-muted-foreground hover:text-foreground transition-colors'
+                )}
+                aria-label="Toggle chat history"
+              >
+                <Menu className="h-5 w-5" />
+              </button>
 
-          {/* Model selector */}
-          <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2">
-            <div className="px-2 py-1 rounded-md border border-[hsla(var(--model-selector-border)_/_0.5)] hover:bg-[hsla(var(--model-selector-bg)_/_0.8)] transition-colors">
-              <ModelSelector selectedModel={selectedModel} onSelectModelAction={setSelectedModel} />
+              {/* Welcome message */}
+              <WelcomeMessage />
+            </div>
+
+            {/* Right controls */}
+            <div className="flex items-center gap-3">
+              {/* Theme toggle button */}
+              <ThemeToggle />
+
+              {/* New chat button */}
+              <button
+                onClick={handleNewChat}
+                className="text-muted-foreground hover:text-foreground"
+                aria-label="New chat"
+              >
+                <PlusSquare className="h-5 w-5" />
+              </button>
             </div>
           </div>
 
-          {/* New chat button */}
-          <button
-            onClick={handleNewChat}
-            className="absolute right-3.5 top-1/2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-            aria-label="New chat"
-          >
-            <PlusSquare className="h-5 w-5" />
-          </button>
-
-          {/* Theme toggle button */}
-          <div className="absolute right-14 top-1/2 -translate-y-1/2">
-            <ThemeToggle />
-          </div>
+          {/* Model selector - desktop: centered, mobile: second row */}
+          {isMobile ? (
+            <div className="w-full flex justify-center mt-2">
+              <div className="px-2 py-1 rounded-md border border-[hsla(var(--model-selector-border)_/_0.5)] hover:bg-[hsla(var(--model-selector-bg)_/_0.8)] transition-colors">
+                <ModelSelector
+                  selectedModel={selectedModel}
+                  onSelectModelAction={setSelectedModel}
+                />
+              </div>
+            </div>
+          ) : (
+            <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2">
+              <div className="px-2 py-1 rounded-md border border-[hsla(var(--model-selector-border)_/_0.5)] hover:bg-[hsla(var(--model-selector-bg)_/_0.8)] transition-colors">
+                <ModelSelector
+                  selectedModel={selectedModel}
+                  onSelectModelAction={setSelectedModel}
+                />
+              </div>
+            </div>
+          )}
         </header>
 
         {/* Chat Area */}
@@ -173,6 +206,7 @@ export default function PageClient() {
                   setInputValue('');
                 }
               }}
+              onResendLastMessageAction={hasUserMessages ? handleResendLastMessage : undefined}
             />
           </div>
         </div>
